requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
# apache-airflow-providers-mysql==5.1.0
apache-airflow-providers-microsoft-mssql